requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
